+++ LogAlong/SelectAccountTableViewController.swift
@@ -30,7 +30,7 @@ class SelectAccountTableViewController: UITableViewController {
     
     @IBAction func okButtonPressed(_ sender: UIButton) {
         
-        delegate?.passIntBack(self, myInt: Float(myIndexPath))
+        delegate?.passIntBack(self, myInt: Double(myIndexPath))
         
         self.dismiss(animated: true, completion: nil)
  /*       let myVC = storyboard?.instantiateViewController(withIdentifier: "testID") as! AddTableViewController
